requirements.txt
Django==2.2.16
pytest==5.3.5
pytest-pythonpath==0.7.3
pytest-django==3.8.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
Pillow==7.0.0
requests==2.22.0
djoser==2.1.0
